+++ target/tidy_worker/test/bench_test.c
@@ -23,26 +23,24 @@ int handle_serialize_bench(int c, const char* html)
     CLOCK_LOG
   }
 
-  // html_workspace_t* w = html_init();
+  tidy_map_str_t opt;
+  tidy_map_init(&opt);
 
-  // // const char *html = "<div><p id=p1><p id=p2><p id=p3><a>link</a><p id=p4><p id=p5><p id=p6></div>";
-  // const char *scope = "html";
-  // vec_eterm_t term_array;
-  // eterm_vec_init(&term_array);
-  // serialize(w, html, scope, &term_array);
-  // char* result = eterm_vec_join(&term_array, "|");
-  // // printf("-> %s\n", result);
-  // // if(strcmp(result, "<html><head></head><body><div><p id=\"p1\"></p><p id=\"p2\"></p><p id=\"p3\"><a>link</a></p><p id=\"p4\"></p><p id=\"p5\"></p><p id=\"p6\"></p></div></body></html>") != 0){
-  // //   eterm_vec_destroy(&term_array);
-  // //   free(result);
-  // //   html_destroy(w);
-  // //   TEST_ERROR
-  // //   return 1;
-  // // }
-  // eterm_vec_destroy(&term_array);
-  // free(result);
+  tidy_map_set(&opt, "TidyShowWarnings", "no");
+  tidy_map_set(&opt, "TidyBodyOnly", "yes");
+  // minify
+  tidy_map_set(&opt, "TidyVertSpace", "auto");
+  tidy_map_set(&opt, "TidyIndentSpaces", "0");
 
-  // html_destroy(w);
+  tidy_workspace_t* w = tidy_init(&opt);
+
+  vec_eterm_t term_array;
+  eterm_vec_init(&term_array);
+  parse(w, html, &term_array);
+  eterm_vec_destroy(&term_array);
+
+  tidy_map_deinit(&opt);
+  tidy_destroy(w);
 
   return 0;
 }